% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{has_package}
\alias{has_package}
\title{Check if a package is available}
\usage{
has_package(pkg)
}
\arguments{
\item{pkg}{Package name.}
}
\value{
Logical.
}
\description{
Thin wrapper around \code{\link[=requireNamespace]{requireNamespace()}} to allow mocking in tests.
}
\keyword{internal}
