% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{wrap_connectors_with_locks}
\alias{wrap_connectors_with_locks}
\title{Wrap Connectors with Lock Protection}
\usage{
wrap_connectors_with_locks(obj, study_path)
}
\arguments{
\item{obj}{A connectors object or connector object}

\item{study_path}{Path to the study folder}
}
\value{
The wrapped object
}
\description{
Recursively wraps all ConnectorFS objects with lock protection.
}
\keyword{internal}
