% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock.R
\name{.onLoad}
\alias{.onLoad}
\title{Package onLoad hook}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Library name}

\item{pkgname}{Package name}
}
\description{
Called when the package is loaded. Registers bundled and cached study
folders as locked (in memory) to prevent accidental data modification.
No files are written to disk.
}
\keyword{internal}
