% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock.R
\name{.onLoad}
\alias{.onLoad}
\title{Package onLoad hook}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Library name}

\item{pkgname}{Package name}
}
\description{
Called when the package is loaded. Locks study folders if this is an
installed package (not in development mode). Also registers S3 methods
for connector integration.
}
\keyword{internal}
