% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock.R
\name{is_study_locked}
\alias{is_study_locked}
\title{Check if a study folder is locked}
\usage{
is_study_locked(study_path)
}
\arguments{
\item{study_path}{Path to the study folder}
}
\value{
Logical indicating if the folder is locked
}
\description{
Checks if a lock file exists for a study folder, indicating that the folder
should not be overwritten (typically because the package has been installed).
}
\keyword{internal}
