% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock.R
\name{unlock_study}
\alias{unlock_study}
\title{Unlock a study folder}
\usage{
unlock_study(study_path)
}
\arguments{
\item{study_path}{Path to the study folder}
}
\value{
Logical indicating success
}
\description{
Removes the lock file from a study folder, allowing it to be overwritten.
}
\examples{
\dontrun{
# Unlock a study folder to allow regeneration
unlock_study("inst/exampledata/cdisc_pilot")
}
}
\keyword{internal}
